fixed 3+ player issues on test client. Made x & y float values, cast to integer when sent to client. Much smoother movement. TODO: fix tickrate calculations that cause spikes in movment (could also be caused by pygame inefficiency)

+++ networking/player.cpp
@@ -7,7 +7,7 @@ player::player(){
     y = 0;
 }
 
-player::player(int a, int b){
+player::player(float a, float b){
     x = a;
     y = b;
 }
